refactor: separate setup
